requirements.txt
atomicwrites==1.4.0
attrs==21.4.0
bcrypt==3.2.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.0.4
colorama==0.4.4
Flask==2.0.3
Flask-Bcrypt==1.0.1
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
gunicorn==20.1.0
idna==3.3
iniconfig==1.1.1
itsdangerous==2.1.1
Jinja2==3.0.3
MarkupSafe==2.1.1
mysqlclient==2.1.0
packaging==21.3
pluggy==1.0.0
py==1.11.0
pycparser==2.21
pyparsing==3.0.8
pytest==7.1.2
python-dotenv==0.20.0
requests==2.27.1
six==1.16.0
spotipy==2.19.0
SQLAlchemy==1.4.33
tomli==2.0.1
urllib3==1.26.9
Werkzeug==2.0.3